dev-libs/mpfr: Add overlay with -mfp-rounding-mode=d and mparam.h.
This ebuild adds a custom src/alpha/mparam.h, along with a patch
to src/mparam_h.in to use it for Alpha. As with my GMP ebuild,
I've generated the mparam.h file on my 667 MHz Compaq XP1000.

I've also added the "-mfp-rounding-mode=d" flag, which fixes
the configure test failure that leads MPFR to otherwise add
"-ffloat-store", which isn't as good for performance compared
to telling GCC to use dynamic rounding mode.